back-end/requirements.txt
cryptography==2.4.2
cycler==0.10.0
decorator==4.0.11
Flask==1.0.2
Flask-Cors==3.0.7
futures==3.1.1
gast==0.2.2
gcloud==0.17.0
geojson==2.4.1
googleapis-common-protos==1.5.5
goslate==1.5.1
grpcio==1.21.1
gunicorn==19.9.0
h5py==2.9.0
httplib2==0.12.0
idna==2.8
imageio==2.1.2
itsdangerous==1.1.0
jws==0.1.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.0
matplotlib==3.1.0
mock==3.0.5
numpy==1.15.2
oauth2client==3.0.0
pandas==0.23.4
Pillow==6.0.0
protobuf==3.6.1
PTable==0.9.2
pycparser==2.19
PyDictionary==1.5.2
pyparsing==2.4.0
Pyrebase==3.0.27
python-dateutil==2.7.3
python-jwt==2.0.1
pytz==2018.5
requests==2.20.0
requests-toolbelt==0.7.0
rsa==4.0
scikit-learn==0.20.3
scipy==1.2.1
six==1.11.0
sklearn==0.0
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
tqdm==4.11.2
urllib3==1.24.1
